move relu6 code from modules to functional
