+++ python/oneflow/nn/modules/activation.py
@@ -18,6 +18,7 @@ from typing import Optional
 
 import oneflow as flow
 from oneflow.nn.modules.module import Module
+import oneflow.nn.functional as F
 
 
 class PReLU(Module):
@@ -161,25 +162,11 @@ class ReLU6(Module):
         self.inplace = inplace
 
     def forward(self, x):
-        if self.inplace:
-            warnings.warn("ReLU6 module do not support inplace now")
-        return flow._C.hardtanh(x, min_val=0.0, max_val=6.0)
+        return F.relu6(x, self.inplace)
 
     def extra_repr(self):
         inplace_str = "inplace=True" if self.inplace else ""
         return inplace_str
-
-
-def relu6(input, inplace=False):
-    r"""relu6(input, inplace=False) -> Tensor
-
-    Applies the element-wise function :math:`\text{ReLU6}(x) = \min(\max(0,x), 6)`.
-
-    See :class:`~oneflow.nn.ReLU6` for more details.
-    """
-    if inplace:
-        warnings.warn("nn.functional.relu6 do not support inplace now")
-    return flow._C.hardtanh(input, min_val=0.0, max_val=6.0)
 
 
 class Tanh(Module):
